html/semantics/embedded-content-0/the-img-element/img.complete.html

<!DOCTYPE HTML>
<html>
<!-- Submitted from TestTWF Paris -->
<head>
    <title>DOM img complete Test</title>
    <meta content="text/html; charset=UTF-8" http-equiv="Content-Type">
    
    <link rel="author" title="Anselm Hannemann" href="http://anselm-hannemann.com/" />
    <link rel="help" href="http://www.w3.org/TR/html5/single-page.html#dom-img-complete" />

    <meta name="assert" content="Tests the complete status of the img-element">

    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>

</head>
<body>
    <h1>DOM img.complete Test</h1>

    <img id="imgTestTag">

    <img src="" id="imgTestTag2">

    <img id="imgTestTag3" style="width: 80px; height:auto;">

    <img id="imgTestTag4">

    <script>
        var imageInstance = document.createElement('img');
        imageInstance.style.display = 'none';

        document.body.appendChild(imageInstance);
    </script>

<div id="log"></div>
<script>
function testImgCompleteNoSrc(){
    assert_true(document.getElementById("imgTestTag").complete);
}

function testImgCompleteEmptySrc(){
    assert_true(document.getElementById("imgTestTag2").complete);
}

test(testImgCompleteNoSrc, "img src omitted");
test(testImgCompleteEmptySrc, "img src empty");

// test if set to true after img is completely available     
var t = async_test("async src complete test");

t.step(
    function(){
        //build path to image file
        var path = document.location.pathname;
        var dir = path.substring(path.indexOf('/', 0)+1, path.lastIndexOf('/'));
        var srcpath =  location.protocol + '//' + location.host + '/' + dir + '/resources/3.jpg'

        document.getElementById("imgTestTag3").src = srcpath;
        
        //test if img.complete is set to false if src is changed
        function testImgCompleteChangeSrc(){
            assert_false(document.getElementById("imgTestTag3").complete)
        }
        test(testImgCompleteChangeSrc,"src changed, should be set to false");

        assert_equals(document.getElementById("imgTestTag3").src, srcpath);
    }
);

document.getElementById("imgTestTag3").onload = t.step_func(
    function(){
        assert_true(document.getElementById("imgTestTag3").complete);
        t.done();
    }
);

// test: The final task that is queued by the networking task source once the resource has been fetched has been queued, but has not yet been run, and the img element is not in the broken state
async_test(function(t) {
    imageInstance.src = 'resources/image-1.jar';
    if(imageInstance.complete == false) {       
        imageInstance.src = 'resources/image-1.jpg';      

        imageInstance.onLoad = t.step_func(function() {
            var startTime = Date.now();
            while((Date.now() - startTime) < 2000) {
                //keep spinning
            }
            
            assert_true(imageInstance.complete, 'Image resource has been fetched but not run yet & image is not in the broken state');
        }); 
        t.done();           
    }
},
'IDL attribute complete returns true when image resource has been fetched but not run yet & image is not in broken state');

// test if broken img does not pass
var t2 = async_test("async src broken test");

t2.step(
    function(){
        //build path to image file
        var path = document.location.pathname;
        var dir = path.substring(path.indexOf('/', 0)+1, path.lastIndexOf('/'));
        var srcpath =  location.protocol + '//' + location.host + '/' + dir + '/resources/brokenimg.ogg'

        document.getElementById("imgTestTag4").src = srcpath;
        
        //test if img.complete is set to false if src is changed
        assert_false(document.getElementById("imgTestTag4").complete, "tsrc changed to broken img, should be set to false");
        
        assert_equals(document.getElementById("imgTestTag4").src, srcpath);
    }
);

var img4 = document.getElementById("imgTestTag4");
img4.onload = img4.onerror = t2.step_func(
    function(event){
        assert_equals(event.type, "error");
        assert_true(document.getElementById("imgTestTag4").complete);
        t2.done();
    }
);

</script>
</body>
</html>
